% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bearing}
\alias{bearing}
\title{Compass bearing of a line between 2 points}
\usage{
bearing(point1, point2, fullcirc = TRUE)
}
\arguments{
\item{point1}{A numeric vector holding x and y coordinates of the first
point.}

\item{point2}{A numeric vector holding x and y coordinates of the second
point.}

\item{fullcirc}{A logical value indicating whether output should be an angle
on [0, 2pi) or [-pi, pi).}
}
\value{
If \code{fullcirc} is \code{FALSE}, the function returns a numeric
  value (angle) from the interval [-pi, pi). \cr
  If \code{fullcirc} is
  \code{TRUE}, the function returns a numeric value (angle) from the interval
  [0, 2pi).
}
\description{
The angle between a line between 2 points in 2-D space and the line from
(0,0) to (0,1) is calcualted. In other words, the compass bearing of a line
between 2 points where north is 0. Angles increase in clockwise direction.
}
\examples{
bearing(c(3,5), c(1,4))
bearing(c(3,5), c(1,4), fullcirc = FALSE)
}
