% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R, R/Ccond.R, R/cyl_cubsec.R,
%   R/cyl_quadsec.R, R/cyl_rect_combine.R, R/cyl_rot_combine.R, R/cyl_vonmises.R
\name{cCopula}
\alias{cCopula}
\alias{cCopula,Copula-method}
\alias{cCopula,cyl_cubsec-method}
\alias{cCopula,cyl_quadsec-method}
\alias{cCopula,cyl_rect_combine-method}
\alias{cCopula,cyl_rot_combine-method}
\alias{cCopula,cyl_vonmises-method}
\title{Calculate the conditional Copula}
\usage{
cCopula(u, copula, cond_on = 2, inverse = F, ...)

\S4method{cCopula}{Copula}(u, copula, cond_on, inverse)

\S4method{cCopula}{cyl_cubsec}(u, copula, cond_on = 2, inverse = F)

\S4method{cCopula}{cyl_quadsec}(u, copula, cond_on = 2, inverse = F)

\S4method{cCopula}{cyl_rect_combine}(u, copula, cond_on = 2, inverse = F)

\S4method{cCopula}{cyl_rot_combine}(u, copula, cond_on = 2, inverse = F)

\S4method{cCopula}{cyl_vonmises}(u, copula, cond_on = 2, inverse = F)
}
\arguments{
\item{u}{Matrix of numeric values in I^2, containing as first collumn
the circular (periodic) and as second the linear dimension.}

\item{copula}{A \code{cyl_copula} or 2-dimensional \code{copula} object.}

\item{cond_on}{Column number of u on which the copula is conditioned. E.g if
\code{cond_on=2}, the function calculates for each element in the first column of u
the Copula conditional on the element in the second column.}

\item{inverse}{logical indicating whether the inverse of the coniditional copula is
calculated.}

\item{...}{Additional arguments.}
}
\value{
A vector containing the values of the distribution of the copula at
\code{[u,-cond_on]} conditional on the values of \code{[u,cond_on]}
}
\description{
Calculate the conditional Copula

Condtional copula

Condtional copula

Condtional copula

Condtional copula

Condtional copula
}
\details{
This is a generic that calls the function \code{copula::cCopula}, for 2-dimensional objects
from the \code{copula}-package for which \code{copula::cCopula} is available. Note that the
input arguments and the output of \code{cylcop::cCopula} differ from those of \code{copula::cCopula}.
}
