% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedcauchy.R
\name{check_precision}
\alias{check_precision}
\title{Check precision of \code{wrappedcauchy} functions}
\usage{
check_precision(scale, K)
}
\arguments{
\item{scale}{The numeric value of the parameter tuning the spread of the
density.}

\item{K}{The limit used to approximate the distribution.}
}
\value{
NULL, warning message if the error is larger than 1%.
}
\description{
Check that the error in the numerical approximation is smaller than 1%
}
