% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cop_plot}
\alias{cop_plot}
\title{Make surface plot or heatmap of cdf or pdf of a copula}
\usage{
cop_plot(
  copula,
  type = c("pdf", "cdf"),
  plot_type = "rgl",
  resolution = 50,
  n_gridlines = 11
)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{type}{A string of characters, either "pdf or "cdf".}

\item{plot_type}{A string of characters. Available plottypes are:
"rgl": surface plot,
"plotly": interactive surface plot, or
"ggplot": heatmap}

\item{resolution}{A numerical value. The density or distribution will be calculated at \code{resolution^2} points.}

\item{n_gridlines}{A numerical value giving the number of gridlines drawn in u and v direction.}
}
\value{
The plot.
}
\description{
Make surface plot or heatmap of cdf or pdf of a copula
}
