% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_cyl}
\alias{cor_cyl}
\title{Estimate a rank-based circular-linear correlation-coefficient-like parameter}
\usage{
cor_cyl(theta, x, plot = TRUE)
}
\arguments{
\item{theta}{A numeric vector of angles (measurements of a circular variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{plot}{A logical value whether a plot of the regression line should be dislpayed.}
}
\value{
The function returns a numeric value between 0 and 1, not -1 and 1, hence "correlation-coefficient-LIKE"
}
\description{
Nomenclature afer Solow 1988 but with radians instead of degrees.
}
