% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R, R/cyl_rect_combine.R,
%   R/cyl_rot_combine.R
\docType{class}
\name{cyl_copula-class}
\alias{cyl_copula-class}
\alias{show,cyl_copula-method}
\alias{show,cyl_rect_combine-method}
\alias{show,cyl_rot_combine-method}
\title{An S4 class of bivariate copulas on the cylinder}
\usage{
\S4method{show}{cyl_copula}(object)

\S4method{show}{cyl_rect_combine}(object)

\S4method{show}{cyl_rot_combine}(object)
}
\arguments{
\item{object}{A \code{cyl_rot_combine}-object}
}
\description{
The class \code{cyl_copula} follows somewhat the structure of the class
copula of the \code{copula}-package.
}
\section{Functions}{
\itemize{
\item \code{show,cyl_copula-method}: What is printed

\item \code{show,cyl_rect_combine-method}: What is printed

\item \code{show,cyl_rot_combine-method}: What is printed
}}

\section{Slots}{

\describe{
\item{\code{name}}{A character string holding the name of the copula.}

\item{\code{parameters}}{A numeric vector holding the parameter values.}

\item{\code{param.names}}{A character vector holding the parameter names.}

\item{\code{param.lowbnd}}{A numeric vector holding the lower bounds of the parameters.}

\item{\code{param.upbnd}}{A numeric vector holding the upper bounds of the parameters.}
}}

