% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cubsec.R
\docType{class}
\name{cyl_cubsec-class}
\alias{cyl_cubsec-class}
\title{An S4 class of copulas with cubic sections}
\description{
This class contains bivariate circular-linear copulas with cubic sections in the linear dimension.
They are periodic in the circular dimension, u, and symmetric with respect to u=0.5. I.e. there is
symmetry between positive and negative angles.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string holding the name of the copula.}

\item{\code{parameters}}{A numeric vector holding the parameter values.}

\item{\code{param.names}}{A character vector holding the parameter names.}

\item{\code{param.lowbnd}}{A numeric vector holding the lower bounds of the parameters.}

\item{\code{param.upbnd}}{A numeric vector holding the upper bounds of the parameters.}
}}

\section{Objects from the Class}{

Objects are created by \code{cyl_cubsec()}.
}

\section{Extends}{

Class \code{cyl_cubsec} extends class \code{cyl_copula}.
}

