% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cubsec.R
\name{cyl_cubsec}
\alias{cyl_cubsec}
\title{Construction of \code{cyl_cubsec} objects}
\usage{
cyl_cubsec(a = 1/(2 * pi), b = 1/(2 * pi))
}
\arguments{
\item{a}{The numeric value of the first parameter of the copula. It must be in [- 1 / (2 * pi)), 1 / (2 * pi))]}

\item{b}{The numeric value of the second parameter of the copula. It must be in [- 1 / (2 * pi)), 1 / (2 * pi))]}
}
\description{
Construction of \code{cyl_cubsec} objects
}
\examples{
cyl_cubsec(a = 0.1, b = -0.1)

}
