% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_gauss.R
\docType{class}
\name{cyl_gauss-class}
\alias{cyl_gauss-class}
\title{An S4 class of bivariate Gaussian copulae}
\description{
Could use \code{normalCopula} from \code{copula}-package instead, just here for legacy
reasons, consistency in output format etc. It is not in any way periodic, just the
usual Gaussian copula.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string holding the name of the copula.}

\item{\code{parameters}}{A numeric vector holding the parameter values.}

\item{\code{param.names}}{A character vector holding the parameter names.}

\item{\code{param.lowbnd}}{A numeric vector holding the lower bounds of the parameters.}

\item{\code{param.upbnd}}{A numeric vector holding the upper bounds of the parameters.}
}}

\section{Objects from the Class}{

Objects are created by \code{cyl_gauss()}.
}

\section{Extends}{

Class \code{cyl_gauss} extends class \code{cyl_copula}.
}

