% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_gauss.R
\name{cyl_gauss}
\alias{cyl_gauss}
\title{Construction of \code{cyl_gauss} objects}
\usage{
cyl_gauss(rho = 0.9)
}
\arguments{
\item{rho}{A numeric value giving the correlation of the copula}
}
\description{
Construction of \code{cyl_gauss} objects
}
\examples{
cyl_gauss(rho=0.5)
}
