% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_quadsec.R
\name{cyl_quadsec}
\alias{cyl_quadsec}
\title{Construction of \code{cyl_quadsec} objects}
\usage{
cyl_quadsec(a = 1/(2 * pi))
}
\arguments{
\item{a}{The numeric value of the parameter of the copula. It must be in [- 1 / (2 * pi)), 1 / (2 * pi))]}
}
\description{
Construction of \code{cyl_quadsec} objects
}
\examples{
cyl_quadsec(a = 0.1)

}
