% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_rect_combine.R
\docType{class}
\name{cyl_rect_combine-class}
\alias{cyl_rect_combine-class}
\title{An S4 class of circular-linear copulae generated from rectangular patches.}
\description{
This class contains bivariate circular-linear copulae generated from
linear-linear bivariate \code{Copula} objects of the \code{copula}-package or circular-linear
copulae of class \code{cyl_copula}. 2 non-overlapping rectangles are laid over the unit
square, both have width 1 in v-direction. In the area covered by the first rectangle,
the one covered by the second and the area not covered by any rectangle all have different copulae.
Rectangle 2 contains the same copula as rectangle 1, but 90 degrees rotated. The copula regions
are combined in a way that the overall result on the entire unit square is lso a copula.
See text for more rigorous explantion.
}
\details{
With appropriate choices of the rectangles (again, see text) this results in copulae
that are periodic in u-direction (and not in v-direction) and therefore are
circular-linear.

When the 2 rectangles are each other's mirror image with respect to u=0.5,
the resulting overall copula is symmetric with respect to u=0.5. I.e. there is
symmetry between positive and negative angles.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string holding the name of the copula.}

\item{\code{parameters}}{A numeric vector holding the parameter values.}

\item{\code{param.names}}{A character vector holding the parameter names.}

\item{\code{param.lowbnd}}{A numeric vector holding the lower bounds of the
parameters.}

\item{\code{param.upbnd}}{A numeric vector holding the upper bounds of the
parameters.}

\item{\code{sym.cop}}{A \code{cyl_vonmises} or \code{Copula} object. The copula in
the rectangles.}

\item{\code{background.cop}}{A \code{cyl_copula} or \code{Copula} object. The copula
where no rectangles overlay the unit square.}

\item{\code{flip_up}}{Logical value indicating whether the copula (\code{sym.cop}) is
rotated 90 degrees in the upper or lower rectangle.}

\item{\code{sym_rect}}{Logical value indicating whether the upper rectangle was
forced to be a mirror image of the lower one with respect to u=0.5 at the
construction of the object.}
}}

\section{Objects from the Class}{
 Objects are created by
  \code{cyl_rect_combine()}.
}

\section{Extends}{

Class \code{cyl_rect_combine} extends class \code{cyl_copula}.
}

