% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_rect_combine.R
\name{cyl_rect_combine}
\alias{cyl_rect_combine}
\title{Construction of \code{cyl_rect_combine} objects}
\usage{
cyl_rect_combine(
  copula,
  background = indepCopula(),
  low_rect = c(0, 0.5),
  up_rect = "symmetric",
  flip_up = TRUE
)
}
\arguments{
\item{copula}{A \code{cyl_vonmises} or \code{Copula} object. The copula in
the rectangles.}

\item{background}{A \code{cyl_copula} or \code{Copula} object. The copula
where no rectangles overlay the unit square.}

\item{low_rect}{A numeric vector containing the lower and upper edge
(u-value) of the lower rectangle.}

\item{up_rect}{A numeric vector containing the lower and upper edge (u-value)
of the upper rectangle,  or the character string "symmetric" if it should be the mirror image
(with respect to u=0.5) of the lower rectangle.}

\item{flip_up}{Logical value indicating whether the copula (\code{sym.cop})
is rotated 90 degrees in the upper or lower rectangle}
}
\description{
Construction of \code{cyl_rect_combine} objects
}
\examples{
cyl_rect_combine(copula = copula::frankCopula(param = 3), background =
cylcop::cyl_vonmises(), low_rect = c(0.1,0.4), up_rect = "symmetric", flip_up
= TRUE)

}
