% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_rot_combine.R
\name{cyl_rot_combine}
\alias{cyl_rot_combine}
\title{Construction of \code{cyl_quadsec} objects}
\usage{
cyl_rot_combine(copula, shift = FALSE)
}
\arguments{
\item{copula}{A linear-linear \code{Copula} object from the \code{copula}-package.}

\item{shift}{A logical indicating whether the (u-periodic) copula shoud be shifted by 0.5 in u direction.}
}
\description{
Construction of \code{cyl_quadsec} objects
}
\examples{
cyl_rot_combine(copula = copula::frankCopula(param = 3), shift = TRUE)
cyl_rot_combine(copula = copula::claytonCopula(param = 10), shift = FALSE)

}
