% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_vonmises.R
\name{cyl_vonmises}
\alias{cyl_vonmises}
\title{Construction of \code{cyl_vonmises} objects}
\usage{
cyl_vonmises(mu = 0, kappa = 1, flip = FALSE)
}
\arguments{
\item{mu}{A numeric value giving the mean of the vonMises function used to construct the copula.}

\item{kappa}{A numeric value giving the concentration of the vonMises function used to construct the copula.}

\item{flip}{A logical indicating whether the copula should be rotated 90 degrees to have negative correlation.}
}
\description{
Construction of \code{cyl_vonmises} objects
}
\examples{
cyl_vonmises(mu=pi, kappa=10, flip = TRUE)

}
