% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{cylcop-deprecated}
\alias{cylcop-deprecated}
\alias{get_inp_lst}
\alias{fit_steplength}
\alias{fit_angle}
\alias{scat_plot}
\alias{traj_plot}
\alias{circ_plot}
\alias{cop_scat_plot}
\alias{cop_plot}
\alias{make_traj}
\alias{qmixedvonmises}
\alias{mle.mixedvonmises}
\title{Deprecated functions}
\usage{
get_inp_lst(mc, defaults, non_default_names)
}
\arguments{
\item{...}{Parameters to be passed to the new versions of the functions}
}
\description{
These functions are provided for compatibility with older version of
the cylcop package. They may eventually be completely
}
\section{Details}{

\tabular{rl}{
  \code{fit_angle()} \tab is replaced by \code{\link{fit_circ_param}()} and \code{\link{fit_circ_np}()}\cr
  \code{fit_steplength()} \tab is replaced by \code{\link{fit_lin_param}()} and \code{\link{fit_lin_np}()}\cr
  \code{scat_plot()} \tab is replaced by \code{\link{plot_joint_scat}()}\cr
  \code{traj_plot()} \tab is replaced by \code{\link{plot_track}()}\cr
  \code{circ_plot()} \tab is replaced by \code{\link{plot_joint_circ}()}\cr
  \code{cop_scat_plot()} \tab is replaced by \code{\link{plot_cop_scat}()}\cr
  \code{cop_plot()} \tab is replaced by \code{\link{plot_cop_surf}()}\cr
  \code{make_traj()} \tab is replaced by \code{\link{traj_sim}()}\cr
  \code{qmixedvonmises()} \tab is replaced by \code{\link{qvonmisesmix}()}\cr
  \code{mle.mixedvonmises()} \tab is replaced by \code{\link{mle.vonmisesmix}()}\cr
}
}

