% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cylcop_get_option}
\alias{cylcop_get_option}
\title{Get options}
\usage{
cylcop_get_option(option = NULL)
}
\arguments{
\item{option}{character string, name of the option}
}
\value{
The value of option
}
\description{
Get options
}
