% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cylcop_set_option}
\alias{cylcop_set_option}
\alias{verbose,}
\alias{silent}
\title{Set Package Options}
\usage{
cylcop_set_option(silent = FALSE)
}
\arguments{
\item{silent}{\link[base]{logical}, suppress all sounds and messages.}
}
\value{
No output, only side effects.
}
\description{
Currently the only option is to toggle verbosity on or off.
}
\examples{
cylcop_set_option(silent = FALSE)
}
\seealso{
\code{\link{cylcop_get_option}()}
}
