% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens}
\alias{dens}
\alias{ddens}
\alias{pdens}
\alias{qdens}
\alias{rdens}
\title{Density, Distribution, Random Number Generation and Quantiles of Kernel Density Estimates}
\usage{
rdens(n, density)

ddens(x, density)

pdens(x, density)

qdens(p, density)
}
\arguments{
\item{n}{\link[base]{integer} value, the number of random samples to be
generated with \code{rdens()}.}

\item{density}{a '\code{\link[stats]{density}}' object (for linear kernel density estimates)
or a '\code{\link[circular]{density.circular}}' object
(for circular kernel density estimates) containing information about the kernel density
  estimate. These objects can be obtained using
  \code{\link{fit_circ_np}()} or
  \code{\link{fit_lin_np}()}.}

\item{x}{\link[base]{numeric} \link[base]{vector} giving the points where
the density or distribution function is evaluated.}

\item{p}{\link[base]{numeric} \link[base]{vector} giving the probabilities where
the quantile function is evaluated.}
}
\value{
\code{ddens()} and \code{pdens()} give a \link[base]{vector} of length \code{length(x)} containing
the density or distribution function at the corresponding values of \code{x}.
\code{qdens()} gives a \link[base]{vector} of length \code{length(p)} containing
the quantiles at the corresponding values of \code{p}. The function \code{rdens()}
generates a \link[base]{vector} of length \code{n} containing the random samples.
}
\description{
Calculate the density (\code{ddens()}), the distribution (\code{pdens()}),
 the quantiles (\code{qdens()}) and generate random
samples (\code{rdens()}) of a kernel density estimate as returned by
\code{\link{fit_circ_np}()} or \code{\link{fit_lin_np}()}.
}
\examples{
set.seed(123)

steps <- rweibull(10, shape=3)
dens <- fit_lin_np(x = steps, limits=c(0,Inf))
ddens(c(0.1,0.3), dens)
pdens(c(0.1,0.3), dens)
qdens(c(0.1,0.3), dens)
rdens(4, dens)

angles <- full2half_circ(
  circular::rvonmises(10, mu = circular::circular(0), kappa = 2)
)
dens <- fit_circ_np(theta = angles)
ddens(c(0.1,0.3), dens)
pdens(c(0.1,0.3), dens)
qdens(c(0.1,0.3), dens)
rdens(4, dens)

}
\seealso{
\code{\link{fit_circ_np}()}, \code{\link{fit_lin_np}()}.
}
