% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{fit_circ_np}
\alias{fit_circ_np}
\title{Fit a Non-Parametric Circular Distribution}
\usage{
fit_circ_np(theta, bandwidth = NULL)
}
\arguments{
\item{theta}{\link[base]{numeric} \link[base]{vector} of angles in \eqn{[-\pi, \pi)}.}

\item{bandwidth}{\link[base]{numeric} value of the kernel density bandwidth.
Default is \code{cylcop::\link{opt_circ_bw}(theta, "nrd")}.}
}
\value{
A '\code{\link[circular]{density.circular}}' object
obtained with the function \code{circular::\link[circular]{density.circular}()}
 of the '\pkg{circular}' package.
}
\description{
This function finds a kernel density estimate using
a von Mises smoothing kernel. It is just a wrapper for
\code{circular::\link[circular]{density.circular}()}.
}
\examples{
set.seed(123)

silent_curr <- cylcop_get_option("silent")
cylcop_set_option(silent = TRUE)

n <- 10 #n (number of samples) is set small for performance.

angles <- rvonmisesmix(n,
  mu = c(0, pi),
  kappa = c(2,1),
  prop = c(0.5, 0.5)
)

bw <- opt_circ_bw(theta = angles,
  method="nrd",
  kappa.est = "trigmoments"
)
dens_non_param <- fit_circ_np(theta = angles,
  parametric = FALSE,
  bandwidth = bw
)

cylcop_set_option(silent = silent_curr)

}
\seealso{
\code{circular::\link[circular]{density.circular}()},
\code{\link{fit_circ_param}()}, \code{\link{fit_lin_np}()},
\code{\link{opt_circ_bw}()}.
}
