% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{fit_circ_param}
\alias{fit_circ_param}
\title{Fit a Circular Univariate Distribution}
\usage{
fit_circ_param(
  theta,
  densfun = c("vonmises", "wrappedcauchy", "vonmisesmix"),
  mu = NULL,
  ncomp = 2
)
}
\arguments{
\item{theta}{\link[base]{numeric} \link[base]{vector} of angles in \eqn{[-\pi, \pi)}.}

\item{densfun}{a \link[base]{character} string describing what distribution
should be fitted (\code{"vonmises"}, \code{"wrappedcauchy"}, or
\code{"vonmisesmix"}).}

\item{mu}{(optional) \link[base]{numeric} \link[base]{vector}, fixed mean direction(s) of the
parametric distribution.}

\item{ncomp}{\link[base]{integer}, number of components of the mixed von Mises distribution.
Only has an effect if \code{densfun="vonmisesmix"}.}
}
\value{
A \link[base]{list} is returned
 containing the estimated parameters, their
standard errors (if available), the log-likelihood,
the AIC and the name of the distribution.
}
\description{
This function finds maximum likelihood
parameter estimates of the marginal circular
distribution (with potentially fixed mean).
}
\examples{
set.seed(123)

silent_curr <- cylcop_get_option("silent")
cylcop_set_option(silent = TRUE)

n <- 10 #n (number of samples) is set small for performance.

angles <- rvonmisesmix(n,
  mu = c(0, pi),
  kappa = c(2,1),
  prop = c(0.5, 0.5)
)

param_estimate <- fit_circ_param(theta = angles,
  densfun = "vonmisesmix",
  ncomp = 2
)
param_estimate_fixed_mean <- fit_circ_param(theta = angles,
  densfun = "vonmisesmix",
  mu = c(0, pi),
  ncomp = 2
)

param_estimate_univ <- fit_circ_param(theta = angles,
  densfun = "vonmises",
  mu = 0
)

cylcop_set_option(silent = silent_curr)

}
\seealso{
\code{circular::\link[circular]{mle.wrappedcauchy}()},
\code{\link{mle.vonmisesmix}()}, \code{\link{fit_lin_param}()},
\code{\link{fit_circ_np}()}.
}
