% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{fit_lin_param}
\alias{fit_lin_param}
\title{Fit a Linear Univariate Distribution}
\usage{
fit_lin_param(
  x,
  densfun = c("beta", "cauchy", "chi-squared", "chisq", "exponential", "exp", "gamma",
    "lognormal", "lnorm", "lognorm", "logistic", "normal", "t", "weibull", "normalmix",
    "weibullmix", "gammamix", "lnormmix"),
  start = NULL,
  ncomp = 2
)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector} of measurements of a linear
random variable.}

\item{densfun}{A \link[base]{character} string describing what distribution
should be fitted (\code{"beta"}, \code{"cauchy"}, \code{"chi-squared"},
\code{"exponential"}, \code{"gamma"}, \code{"lognormal"}, \code{"logistic"},
\code{"normal"}, \code{"t"}, \code{"weibull"},\code{"normalmix"},
\code{"weibullmix"}, \code{"gammamix"}, or \code{"lnormmix"}).}

\item{start}{(optional, except when \code{densfun = "chi-squared"})
named \link[base]{list} containing the parameters to be optimized with initial
values.}

\item{ncomp}{\link[base]{integer}, number of components of the mixed distribution.
Only has an effect if \code{densfun \%in\% c("normalmix", "weibullmix", "gammamix",
"lnormmix")}.}
}
\value{
A \link[base]{list} containing the estimated parameters,
their standard errors,
the log-likelihood, the AIC and the name of the distribution.
}
\description{
This function finds maximum likelihood parameter estimates
of the marginal linear distribution. It uses
\code{MASS::\link[MASS]{fitdistr}()} and
\code{mixR::\link[mixR]{mixfit}()}. For more control, use these functions
directly.
}
\examples{
require(graphics)
set.seed(123)

silent_curr <- cylcop_get_option("silent")
cylcop_set_option(silent = TRUE)

n <- 100 #n (number of samples) is set small for performance.

x <- rweibull(n, shape = 10)

weibull <- fit_lin_param(x = x, densfun = "weibull")
gamma <- fit_lin_param(x = x, densfun = "gamma")
chisq <- fit_lin_param(x = x, densfun = "chi-squared", start = list(df = 1))

true_dens <- dweibull(seq(0, max(x), length.out = 200),
  shape = 10
)
dens_weibull <- dweibull(seq(0, max(x),length.out = 200),
  shape = weibull$coef$shape,
  scale = weibull$coef$scale
)
dens_gamma <- dgamma(seq(0, max(x),length.out = 200),
  shape = gamma$coef$shape,
  rate = gamma$coef$rate
)
dens_chisq <- dchisq(seq(0, max(x),length.out = 200),
  df = chisq$coef$df
)

plot(seq(0,max(x),length.out = 200), true_dens, type = "l")
lines(seq(0,max(x),length.out = 200), dens_weibull, col = "green")
lines(seq(0,max(x),length.out = 200), dens_gamma, col = "blue")
lines(seq(0,max(x),length.out = 200), dens_chisq, col = "cyan")

cylcop_set_option(silent = silent_curr)

}
\seealso{
\code{\link{fit_circ_param}()}, \code{\link{fit_lin_np}()}.
}
