% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_mle.R
\name{fit_LL}
\alias{fit_LL}
\title{Carry out MLE of copula parameters}
\usage{
fit_LL(
  copula,
  emp_cop,
  param_name,
  start,
  lower,
  upper,
  optim.method,
  optim.control,
  estimate.variance,
  traceOpt
)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{emp_cop}{A numeric matrix with 2 columns holding the empirical copula values.}

\item{param_name}{A vector of character strings holding the names of the parameters to be optimized.
These can be any parameters in \code{copula@parameters}.}

\item{start}{A vector of staring values of the parameters.}

\item{lower}{A vector of lower bounds of the parameters.}

\item{upper}{A vector of upper bounds of the parameters.}

\item{optim.method}{optimizer used in \code{optim()}, can be
"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", or "Brent"}

\item{optim.control}{A list of additional controls passed to \code{optim()}.}

\item{estimate.variance}{A logical value, denoting whether to include an
estimate of the variance (NOT YET IMPLEMENTED).}

\item{traceOpt}{A logical value, denoting whether to print information regarding
convergence, current values, etc. during the optimization process.}
}
\value{
A list containing the same type of \code{cyl_copula} object as \code{copula},
 but with optimized parameters, the log likelihood and the AIC.
}
\description{
Carry out MLE of copula parameters
}
