% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{fit_steplength}
\alias{fit_steplength}
\title{Fit linear univariate distribution}
\usage{
fit_steplength(
  x,
  parametric = c("beta", "cauchy", "chi-squared", "exponential", "gamma", "lognormal",
    "logistic", "normal", "t", "weibull", FALSE),
  start = NULL,
  bandwidth = NULL
)
}
\arguments{
\item{x}{A numeric vector of measurements of a linear random variable.}

\item{parametric}{Either a character string describing what (continuous)
distribution should be fitted or FALSE if a non-parametric estimation
(kernel density) should be made.}

\item{start}{OPTIONAL: A numeric vector holding the starting values  for the ML parameter estimation.}

\item{bandwidth}{The numeric value of for the kernel density bandwidth.
Default is  \code{cylcop::opt_lin_bw(x, "adhoc")}.}
}
\value{
If a parametric estimate is made, a list containing the estimated parameters, their
standard errors and the loglikelihood is returned.
If a nonparametric estimate is made, the output comes from
the function \code{GoFKernel::density.reflected()}.
}
\description{
This function finds parameter estimates of the marginal linear
distribution, or gives a kernel density estimate.
}
