% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full2half_circ}
\alias{full2half_circ}
\title{Convert angle from full circle to half circle}
\usage{
full2half_circ(angle)
}
\arguments{
\item{angle}{A numeric vector containing angles or \code{\link{circular}}-objects in [0, 2pi).}
}
\value{
A numeric vector containing the angles in [-pi, pi).
}
\description{
Converts angle from the interval [0, 2pi) to [-pi, pi).
}
\examples{
full2half_circ(1.5*pi)
}
