% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_marg}
\alias{get_marg}
\title{Get univariate distributions}
\usage{
get_marg(marg_type)
}
\arguments{
\item{marg_type}{A character string corresponding to the type of univariate distribution, e.g. "norm"}
}
\value{
Give A list of functions of a univariate distribution (random gneration, density, distribution, quantiles)
}
\description{
Get univariate distributions
}
\examples{
get_marg("norm")
}
