% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{half2full_circ}
\alias{half2full_circ}
\title{Convert angle from half circle to full circle}
\usage{
half2full_circ(angle)
}
\arguments{
\item{angle}{The numeric value of an angle or a \code{\link{circular}}-objekt in [-pi, pi).}
}
\value{
The numeric value of the angle in [0, 2pi).
}
\description{
Converts angle from the interval [-pi, pi) to [0, 2pi).
}
\examples{
half2full_circ(-0.5*pi)
}
