% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectory.R
\name{make_traj}
\alias{make_traj}
\title{Generate a trajectory}
\usage{
make_traj(
  n,
  copula,
  marginal_circ = c("vonmises", "wrappedcauchy", "mixedvonmises", "dens", "hnorm"),
  parameter_circ = list(0, 1),
  marginal_lin,
  parameter_lin = list()
)
}
\arguments{
\item{n}{Number of trajectory steps to generate.}

\item{copula}{A \code{cyl_copula} object.}

\item{marginal_circ}{A character string denoting the name of the circular
distribution. "vonmises","wrappedcauchy", "dens" (for kernel density
estimate), or "hnorm".}

\item{parameter_circ}{A list of parameters of the circular marginal dsitribution as taken by the functions
\code{qvonmises()},\code{qwrappedcauchy()}, or \code{qhorm()}. For \code{marginal_circ = "dens"}, \code{parameter_circ} must be the density object,
which can be obtained using \code{fit_angle(...,parametric = FALSE)}.}

\item{marginal_lin}{A character string denoting the name of the linear distribution. I.e. the name of its distribution function without the "p"
e.g. "norm" for normal distribution.}

\item{parameter_lin}{A list of parameters of the linear marginal dsitribution. For \code{marginal_circ = "dens"}, \code{parameter_circ} must be the density object,
which can be obtained using \code{fit_steplength(...,parametric = FALSE)}.}
}
\value{
A data.frame containing the trajectory.
}
\description{
Generate a trajectory
}
