% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{mi_binned}
\alias{mi_binned}
\title{Estimate the mutual information between a circular and a linear random
variable}
\usage{
mi_binned(theta, x, normalize = T, symmetrize = F)
}
\arguments{
\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear
variable).}

\item{normalize}{A logical value whether the mutual information should be
normalized to lie within [0,1].}

\item{symmetrize}{A logical value whether it should be assumed that positive
and negative angles are equivalent. If they are indeed exactly equivalent
and \code{symmetrize=F}, the maximal normalized mutual information will be about
0.5.}
}
\value{
A numeric value, the mutual information between \code{theta} and \code{x}.
}
\description{
Estimate the mutual information between a circular and a linear random
variable
}
