% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedvonmises.R
\name{mle.mixedvonmises}
\alias{mle.mixedvonmises}
\title{mixed von Mises Maximum Likelihood Estimates}
\usage{
mle.mixedvonmises(theta, mu = NULL)
}
\arguments{
\item{theta}{vector of angles}

\item{mu}{(optional) vector holding the 2 mean directions (angles). If not specified
the mean directions are estimated.}
}
\value{
A list containing the optimized parameters \code{mu}, \code{kappa}, and \code{prop}.
}
\description{
Computes the maximum likelihood estimates for the parameters of a von Mises distribution:
the mean directions, the concentration parameters, and the proportion of the 2 distributions.
}
\details{
Code is a simplified version of \code{movMF::movMF()} with the added
feature of optionally fixed mean directions, see references.
}
\references{
Hornik, K. and Grun, B. (2014). movMF: An R Package for Fitting Mixtures of von Mises-Fisher Distributions. _J. Stat. Softw., 58(10)_, 164-168.
}
