% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ccond.R
\name{numerical_conditional_cop}
\alias{numerical_conditional_cop}
\title{Numerically calculate the conditional copula}
\usage{
numerical_conditional_cop(u, copula, cond_on)
}
\arguments{
\item{u}{Matrix of numeric values in I^2, containing as first collumn
the circular (periodic) and as second the linear dimension.}

\item{copula}{A \code{cyl_copula} or 2-dimensional \code{copula} object.}

\item{cond_on}{Column number of u on which the copula is conditioned. E.g if
\code{cond_on=2}, the function calculates for each element in the first column of u
the Copula conditional on the element in the second column.}
}
\value{
A vector containing the values of the distribution of the copula at
\code{[u,-cond_on]} conditional on the values of \code{[u,cond_on]}
}
\description{
Numerically calculate the conditional copula
}
