% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_auto.R
\name{opt_auto}
\alias{opt_auto}
\title{Black-box to find best fitting copula}
\usage{
opt_auto(theta, x)
}
\arguments{
\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}
}
\value{
A list containing 3 lists: Descriptions of the copulae, the
\code{cyl_copula}-objects with fitted parameters, the AIC.
}
\description{
The parameters of 15 different circular-linear copulas are fitted and sorted
according to AIC. For each copula, first, a starting value for the maximum
likelihood estimation (MLE) is found using \code{cylcop::optCor()} or
\code{cylcop::optTau()}. Then MLE with a "reasonable" setup is carried out.
If MLE fails, parameters obtained with \code{cylcop::optCor()} or
\code{cylcop::optTau()} are reported.
}
