% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{opt_lin_bw}
\alias{opt_lin_bw}
\title{Find the optimal bandwidth for linear kernel density estimate}
\usage{
opt_lin_bw(x, loss = c("KullbackLeibler", "adhoc"))
}
\arguments{
\item{x}{A numeric vector of linear measurements.}

\item{loss}{A string of characters describing the loss function, either "KullbackLeibler" (leading to an MLE estimate),
or "adhoc" leading to a rule-of-thumb estimate.}
}
\value{
The numeric value of the optimized bandwidth.
}
\description{
Find the optimal bandwidth for linear kernel density estimate
}
