% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{optCor,cyl_cubsec-method}
\alias{optCor,cyl_cubsec-method}
\title{Roughly estimate cyl_cubsec parameters from some correlation metric}
\usage{
\S4method{optCor}{cyl_cubsec}(
  copula,
  theta,
  x,
  acc,
  n,
  method = c("cor_cyl", "mi"),
  parameter = "both"
)
}
\arguments{
\item{copula}{A \code{cyl_cubsec} object}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{acc}{A numeric value, the interval of a and/or b at which to evaluate
the correlation. DEFAULT: 0.01}

\item{n}{The number of sample points at each optimization step (value of a
and/or b). DEFAULT: 10000}

\item{method}{A string of characters, describing what correlation metric to
use. Either a rank-based circular-linear coefficient ("cor_cyl") or mutual
information ("mi").}

\item{parameter}{A character string specifying which parameter to optimize,
"a", "b", or "both"}
}
\value{
The estimated value of a, or b, or a numeric vector holding the
  estimated values of a and b.
}
\description{
Roughly estimate cyl_cubsec parameters from some correlation metric
}
