% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{optCor,cyl_quadsec-method}
\alias{optCor,cyl_quadsec-method}
\title{Roughly estimate cyl_quadsec parameter from some correlation metric}
\usage{
\S4method{optCor}{cyl_quadsec}(copula, theta, x, acc, n, method = c("cor_cyl", "mi"))
}
\arguments{
\item{copula}{A \code{cyl_quadsec} object}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{acc}{A numeric value, the interval of a at which to evaluate the
correlation. DEFAULT: 0.01}

\item{n}{The number of sample points at each optimization step (value of a). DEFAULT: 10000}

\item{method}{A string of characters, describing what correlation metric to use.
Either a rank-based circular-linear coefficient ("cor_cyl") or
mutual information ("mi").}
}
\value{
The estimated value of a.
}
\description{
Roughly estimate cyl_quadsec parameter from some correlation metric
}
