% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{optCor,cyl_rect_combine-method}
\alias{optCor,cyl_rect_combine-method}
\title{Roughly estimate cyl_rect_combine parameters from some correlation metric}
\usage{
\S4method{optCor}{cyl_rect_combine}(
  copula,
  theta,
  x,
  acc,
  n,
  method = c("tau", "cor_cyl", "mi_cyl"),
  background = FALSE
)
}
\arguments{
\item{copula}{A \code{cyl_rect_combine} object}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{acc}{A numeric value, the interval of the copula parameter at which to evaluate the
correlation. DEFAULT: 0.5}

\item{n}{The number of sample points at each optimization step (parameter value). DEFAULT: 10000}

\item{method}{A string of characters, describing what correlation metric to
use. Either Kendall's tau ("tau"), a rank-based circular-linear coefficient
("cor_cyl"), or mutual information ("mi_cyl").}

\item{background}{A logical value describing whether to optimize the parameter of the background copula, (TRUE) or
the one of the copula in the rectangles(FALSE).}
}
\value{
The estimated value of the copula parameter.
}
\description{
Roughly estimate cyl_rect_combine parameters from some correlation metric
}
