% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{optCor,cyl_vonmises-method}
\alias{optCor,cyl_vonmises-method}
\title{Roughly estimate cyl_vonmises parameter from some correlation metric}
\usage{
\S4method{optCor}{cyl_vonmises}(copula, theta, x, acc, n, method = c("cor_cyl", "mi"))
}
\arguments{
\item{copula}{A \code{cyl_vonmises} object}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{acc}{A numeric value, the interval of kappa at which to evaluate the
correlation. DEFAULT: 0.5}

\item{n}{The number of sample points at each optimization step (kappa value). DEFAULT: 10000}

\item{method}{A string of characters, describing what correlation metric to use.
Either a rank-based circular-linear coefficient ("cor_cyl") or
mutual information ("mi").}
}
\value{
The estimated value of kappa.
}
\description{
It only makes sense to optimize kappa, mu does not influence correlation
}
