% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_mle.R
\name{optML}
\alias{optML}
\title{Estimate cyl_copula parameters according to maximum likelihood}
\usage{
optML(
  copula,
  theta,
  x,
  parameters,
  start,
  lower = NULL,
  upper = NULL,
  optim.method = "L-BFGS-B",
  optim.control = list(maxit = 100),
  estimate.variance = F,
  traceOpt = F
)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{parameters}{A vector of character strings holding the names of the parameters to be optimized.
These can be any parameters in \code{copula@parameters}.}

\item{start}{A vector of staring values of the parameters.}

\item{lower}{OPTIONAL: A vector of lower bounds of the parameters.}

\item{upper}{OPTIONAL: A vector of upper bounds of the parameters.}

\item{optim.method}{optimizer used in \code{optim()}, can be
"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", or "Brent"}

\item{optim.control}{A list of additional controls passed to \code{optim()}.}

\item{estimate.variance}{A logical valu, denoting whether to include an
estimate of the variance (NOT YET IMPLEMENTED).}

\item{traceOpt}{A logical value, denoting whether to print information regarding
convergence, current values, etc. during the optimization process.}
}
\value{
A list containing the same type of \code{cyl_copula} object as \code{copula},
 but with optimized parameters, the log likelihood and the AIC.
}
\description{
The code of this function is based on \code{copula::fitCopula()}.
}
