% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{optTau}
\alias{optTau}
\alias{optTau,cyl_rect_combine-method}
\alias{optTau,Copula-method}
\title{Estimate copula parameters based on Kendall's tau of the empirical copula}
\usage{
optTau(copula, theta, x)

\S4method{optTau}{cyl_rect_combine}(copula, theta, x)

\S4method{optTau}{Copula}(copula, theta, x)
}
\arguments{
\item{copula}{A \code{Copula} object (from the copula package) or a
\code{cyl_rect_combine} object.}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}
}
\value{
A vector holding the parameter estimates.
}
\description{
Estimate copula parameters based on Kendall's tau of the empirical copula
}
\section{Methods (by class)}{
\itemize{
\item \code{cyl_rect_combine}: Works only for rectangles spanning the entire unit square, see text.

\item \code{Copula}: Only here for consistency, just calls function from copula package.
}}

