% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R
\name{param_num_checked}
\alias{param_num_checked}
\title{Check new \code{cyl_copula} parameters}
\usage{
param_num_checked(copula, param_val, param_name)
}
\arguments{
\item{copula}{A \code{cyl_copula}-object}

\item{param_val}{A numeric vector holding the values to which you want to
change certain parameters stored in \code{copula@parameters}.}

\item{param_name}{A vector vector of character strings holding the names of the parameters to be
changed.}
}
\value{
The function returns a numeric vector holding the indices of
  \code{param_val} in \code{copula@parameters}.
}
\description{
Check new \code{cyl_copula} parameters
}
