% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_mle.R
\name{prep_n_check_ML}
\alias{prep_n_check_ML}
\title{Check input and prepare data for MLE}
\usage{
prep_n_check_ML(copula, theta, x, param_name, start, lower, upper)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear variable).}

\item{param_name}{A vector of character strings holding the names of the parameters to be optimized.
These can be any parameters in \code{copula@parameters}.}

\item{start}{A vector of staring values of the parameters.}

\item{lower}{OPTIONAL: A vector of lower bounds of the parameters.}

\item{upper}{OPTIONAL: A vector of upper bounds of the parameters.}
}
\value{
The function returns a list containing the empirical copula of the data and
reasonable upper and lower bounds (if not specified in the input). It also checks the input.
}
\description{
Check input and prepare data for MLE
}
