% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_param}
\alias{print_param}
\title{Print arguments of a function call}
\usage{
print_param(
  fun,
  params = list(),
  start = 2,
  param_print = seq(start, length(formals(fun)))
)
}
\arguments{
\item{fun}{The function.}

\item{params}{A list of characters holding the arguments passed to that function (named or unnamed).}

\item{start}{An integer value, indicating with which function argument printing starts.}

\item{param_print}{An integer value, indicating how many arguments are printed.}
}
\value{
A string consisting of parameter names and values with wich \code{fun} was called
}
\description{
Print arguments of a function call
}
\examples{
print_param(dnorm, list(sd=1),start=1)
}
