% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R
\name{printCop}
\alias{printCop}
\alias{printCop,cyl_copula-method}
\alias{printCop,Copula-method}
\title{Print copula parameters}
\usage{
printCop(copula)

\S4method{printCop}{cyl_copula}(copula)

\S4method{printCop}{Copula}(copula)
}
\arguments{
\item{copula}{A \code{cyl_copula} or \code{Copula} object.}
}
\value{
NULL, prints the first 3 parameters of \code{copula} and their values
 in the console.
}
\description{
Used e.g. for printing a short summary of the copula in \code{make_traj()} and other
functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{cyl_copula}: Print \code{cyl_copula} parameters

\item \code{Copula}: Print \code{Copula} parameters
}}

