% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R
\name{prob,cyl_copula-method}
\alias{prob,cyl_copula-method}
\title{Calculate the C-volume of a bivariate copula}
\usage{
\S4method{prob}{cyl_copula}(x, l, u)
}
\arguments{
\item{x}{A \code{cyl_copula} object}

\item{l}{A numeric vector of length 2 holding the coordinates of the lower left corner in I^2}

\item{u}{A numeric vector of length 2 holding the coordinates of the upper right corner in I^2}
}
\description{
This is a method corresponding to the generic \code{prob} in the
\code{copula}-package
}
