% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedvonmises.R
\name{qmixedvonmises}
\alias{qmixedvonmises}
\title{Quantiles of the mixed vonMises distribution}
\usage{
qmixedvonmises(p, mu1, mu2, kappa1, kappa2, prop)
}
\arguments{
\item{p}{vector of probabilities}

\item{mu1}{mean of the first component}

\item{mu2}{mean of the second component}

\item{kappa1}{concentration parameter of the first component}

\item{kappa2}{concentration parameter of the first component}

\item{prop}{mixing proportion}
}
\value{
quantiles of the mixed vonMises distribution.
}
\description{
Quantiles of the mixed vonMises distribution
}
