% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{scat_plot}
\alias{scat_plot}
\title{Make scatterplot of turning angles ans steplengths}
\usage{
scat_plot(traj, periodic = FALSE)
}
\arguments{
\item{traj}{A data.frame containing the trajectory. Must contain collumns \code{traj$angle} and \code{traj$steplength}.}

\item{periodic}{A logical value denoting whether the plot should be periodically extended past -pi and pi.}
}
\value{
The scatterplot
}
\description{
Make scatterplot of turning angles ans steplengths
}
