% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cop_corr.R
\name{search_cor}
\alias{search_cor}
\title{Finds the value of a specified parameter which gives correlation closest to empirical copula}
\usage{
search_cor(
  copula,
  data,
  method = c("cor_cyl", "mi_cyl"),
  acc,
  n,
  min,
  max,
  param_name
)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{data}{A data.frame containing circular and linear measurements.}

\item{method}{A string of characters, describing what correlation metric to
use. Either a rank-based circular-linear coefficient ("cor_cyl") or mutual
information ("mi_cyl").}

\item{acc}{A numeric value, the interval of the copula parameter at which to evaluate the
correlation.}

\item{n}{The number of sample points at each optimization step (parameter value).}

\item{min}{The minimum value of parameter to test.}

\item{max}{The maximum value of parameter to test.}

\item{param_name}{A character string, the name of the parameter to be varied.}
}
\value{
The value of parameter \code{param_name} that gives correlation closest to the empirical copula
of the data in \code{data}.
}
\description{
This function assumes that correlation and MI increase monotonously with the copula parameter
}
