% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R, R/cyl_cubsec.R,
%   R/cyl_quadsec.R, R/cyl_rect_combine.R, R/cyl_rot_combine.R, R/cyl_vonmises.R
\name{setCopParam}
\alias{setCopParam}
\alias{setCopParam,cyl_cubsec-method}
\alias{setCopParam,cyl_quadsec-method}
\alias{setCopParam,cyl_rect_combine-method}
\alias{setCopParam,cyl_rot_combine-method}
\alias{setCopParam,cyl_vonmises-method}
\title{Change attributes of \code{cyl_copula} objects}
\usage{
setCopParam(copula, param_val, param_name = NULL, ...)

\S4method{setCopParam}{cyl_cubsec}(copula, param_val, param_name)

\S4method{setCopParam}{cyl_quadsec}(copula, param_val, param_name)

\S4method{setCopParam}{cyl_rect_combine}(copula, param_val, param_name)

\S4method{setCopParam}{cyl_rot_combine}(copula, param_val, param_name)

\S4method{setCopParam}{cyl_vonmises}(copula, param_val, param_name)
}
\arguments{
\item{copula}{A \code{cyl_copula} object.}

\item{param_val}{A numeric vector holding the values to which you want to
change certain parameters stored in \code{copula@parameters}.}

\item{param_name}{A vector vector of character strings holding the names of the parameters to be
changed.}

\item{...}{Additional arguments.}
}
\value{
A \code{cyl_copula}-object with the changed parameters.
}
\description{
Change attributes of \code{cyl_copula} objects
}
