% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cop_class.R, R/cyl_rect_combine.R,
%   R/cyl_rot_combine.R
\name{show,cyl_copula-method}
\alias{show,cyl_copula-method}
\alias{show,cyl_rect_combine-method}
\alias{show,cyl_rot_combine-method}
\title{Print Information of '\code{cyl_copula}' Objects}
\usage{
\S4method{show}{cyl_copula}(object)

\S4method{show}{cyl_rect_combine}(object)

\S4method{show}{cyl_rot_combine}(object)
}
\arguments{
\item{object}{\R object of class '\code{\linkS4class{cyl_copula}}'.}
}
\value{
An invisible \code{NULL}. As side effect, information on \code{object} is
printed.
}
\description{
Methods for function \code{\link[methods]{show}()} in package \pkg{cylcop}
}
