% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{traj_plot}
\alias{traj_plot}
\title{Plot the trajectories locations in x-y space}
\usage{
traj_plot(traj)
}
\arguments{
\item{traj}{A data.framecontaining the trajectory. Must contain collumns \code{traj$angle} and \code{traj$steplength}.}
}
\value{
The plot
}
\description{
Plot the trajectories locations in x-y space
}
