% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{circ_plot}
\alias{circ_plot}
\title{Make circular scatterplot of turning angles an steplengths}
\usage{
circ_plot(traj)
}
\arguments{
\item{traj}{A data.framecontaining the trajectory. Must contain collumns \code{traj$angle} and \code{traj$steplength}.}
}
\value{
the plot.
}
\description{
Make circular scatterplot of turning angles an steplengths
}
