% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cop_scat_plot}
\alias{cop_scat_plot}
\title{Make scatterplot of copula values}
\usage{
cop_scat_plot(traj)
}
\arguments{
\item{traj}{#' @param traj A data.framecontaining the trajectory. Must contain collumns \code{traj$angle} and \code{traj$steplength}.}
}
\value{
The scatterplot.
}
\description{
Make scatterplot of copula values
}
