% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cop_scat_plot}
\alias{cop_scat_plot}
\title{Make scatterplot of copula values}
\usage{
cop_scat_plot(input)
}
\arguments{
\item{input}{Either
a data.frame containing the trajectory. Must contain collumns
\code{traj$cop_u} and \code{traj$cop_v}\\
or\\
a \code{copula} object or \code{cyl_copula} object.}
}
\value{
The scatterplot.
}
\description{
Make scatterplot of copula values
}
