% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylcop-package.R
\docType{package}
\name{cylcop-package}
\alias{cylcop}
\alias{cylcop-package}
\title{cylcop: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Hodel \email{florian.hodel@yahoo.com} (\href{https://orcid.org/0000-0002-0099-1006}{ORCID})

}
\keyword{internal}
