% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylcop-package.R
\docType{package}
\name{cylcop-package}
\alias{cylcop}
\alias{cylcop-package}
\title{cylcop: Circular-Linear Copulas with Angular Symmetry for Movement Data}
\description{
Classes (S4) of circular-linear, symmetric copulas with corresponding
    methods, extending the 'copula' package. These copulas are especially useful
    for modelling correlation in discrete-time movement data. Methods for density, 
    (conditional) distribution, random number generation, bivariate dependence
    measures and fitting parameters using maximum likelihood and other 
    approaches. The package also contains methods for visualizing movement data
    and copulas.
}
\author{
\strong{Maintainer}: Florian Hodel \email{florian.hodel@yahoo.com} (\href{https://orcid.org/0000-0002-0099-1006}{ORCID})

}
\keyword{internal}
