% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cylcop_get_option}
\alias{cylcop_get_option}
\title{Get Package Options}
\usage{
cylcop_get_option(option = NULL)
}
\arguments{
\item{option}{\link[base]{character} string, the name of the option.}
}
\value{
The \link[base]{numeric} value of option. If no argument is provided, a list
of all options is printed.
}
\description{
Currently the only option (\code{"silent"}) is to toggle verbosity on or off.
}
\examples{
cylcop_get_option("silent")
cylcop_get_option()
}
\seealso{
\code{\link{cylcop_set_option}()}
}
