% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens}
\alias{dens}
\alias{rdens}
\alias{ddens}
\alias{pdens}
\alias{qdens}
\title{Density, distribution, random number generation and quantiles for a kernel density estimate}
\usage{
rdens(n, density)

ddens(x, density)

pdens(x, density)

qdens(p, density)
}
\arguments{
\item{n}{Number of random samples to be generated.}

\item{density}{A list containing information about the kernel density
estimate. The structure of the list must be as returned by \code{fit_angle()} or
\code{fit_steplength()}.}

\item{x}{A numeric vector of measurements.}

\item{p}{A numeric vector of probabilities.}
}
\description{
Density, distribution, random number generation and quantiles for a kernel density estimate
}
