% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{fit_angle}
\alias{fit_angle}
\title{Fit circular univariate distribution}
\usage{
fit_angle(
  theta,
  parametric = c("vonmises", "wrappedcauchy", FALSE),
  bandwidth = NULL
)
}
\arguments{
\item{theta}{A numeric vector of angles in [-pi, pi).}

\item{parametric}{Either a character string describing what distribution
should be fitted "vonmises" or "wrappedcauchy" or FALSE if a non-parametric
estimation (kernel density) should be made.}

\item{bandwidth}{The numeric value of for the kernel density bandwidth.
Default is  \code{cylcop::opt_circ_bw(theta, "adhoc")}.}
}
\value{
If a parametric estimate is made, a list containing the estimated parameters, their
standard errors and the loglikelihood is returned.
If a nonparametric estimate is made, the output comes from
the funtion \code{circular::density.circular()}.
}
\description{
This function finds parameter estimates of the marginal circular
distribution, or gives a kernel density estimate.
}
