% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{mi_cyl}
\alias{mi_cyl}
\title{Estimate the mutual information between a circular and a linear random
variable}
\usage{
mi_cyl(theta, x, normalize = T, symmetrize = F)
}
\arguments{
\item{theta}{A numeric vector of angles (measurements of a circular
variable).}

\item{x}{A numeric vector of steplengths (measurements of a linear
variable).}

\item{normalize}{A logical value whether the mutual information should be
normalized to lie within [0,1].}

\item{symmetrize}{A logical value whether it should be assumed that positive
and negative angles are equivalent.}
}
\value{
A numeric value, the mutual information between \code{theta} and \code{x}.
}
\description{
The mutual information can be normalized to lie between 0 ans 1
by dividing by the product of the entropies of \code{x} and \code{theta}.
Even if \code{x} and \code{theta} are perfectly correlated, the normalized
mutual information will not be 1 if the underlying copula is periodic and
symmetric. Therefore, we can set \code{symmetrize=T} to set all u-values of
the empirical copula that are larger than 0.5 to 1-0.5. The mutual information
is then calculated from those values and is exactly 1 in the case of
perfect correlation as captured by e.g.
\code{cyl_rect_combine(normalCopula(1))}.
}
