% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_margin.R
\name{opt_circ_bw}
\alias{opt_circ_bw}
\title{Find the optimal bandwidth for circular kernel density estimate}
\usage{
opt_circ_bw(theta, loss = c("KullbackLeibler", "adhoc"), kappa.est = "ML")
}
\arguments{
\item{theta}{A numeric vector of angles in [-pi, pi).}

\item{loss}{A string of characters describing the loss function, either "KullbackLeibler" (leading to an MLE estimate),
or "adhoc" leading to a rule-of-thumb estimate.}

\item{kappa.est}{Character string describing how the spread is estimated. Either maximum likelihood "ML"
or trigonometric moment "trigmoments"}
}
\value{
The numeric value of the optimized bandwidth.
}
\description{
The ad-hoc method of finding the bandwidth parameter might give very bad results, especially for multimodal population distributions.
In these cases it might help to set \code{kappa.est=trigmoments}.
}
