% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{wasserstein}
\alias{wasserstein}
\title{Calculate the Wasserstein distance.}
\usage{
wasserstein(copula, copula2 = NULL, data = NULL, n_sample, n_grid = 2500)
}
\arguments{
\item{copula}{\R object of class '\code{\linkS4class{cyl_copula}}'.
or '\code{\linkS4class{Copula}}' (package '\pkg{copula}', only 2-dimensional).}

\item{copula2}{(optional) \R object of class '\code{\linkS4class{cyl_copula}}'.
or '\code{\linkS4class{Copula}}' (package '\pkg{copula}', only 2-dimensional).}

\item{data}{\link[base]{numeric} \link[base]{matrix} with 2 columns, holding the
angles and step lengths}

\item{n_sample}{\link[base]{integer} number of samples to use when comparing two copula PDFs}

\item{n_grid}{number of gridcells}
}
\value{
\link[base]{integer}, the second Wasserstein distance
}
\description{
The second Wasserstein distance, based on the Euclidean distance,
 between two copula PDFs on a grid, or between a copula pdf and pseudo-observations.
}
