% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedcauchy.R
\name{wrappedcauchy}
\alias{wrappedcauchy}
\alias{rwrappedcauchy}
\alias{dwrappedcauchy}
\alias{pwrappedcauchy}
\alias{qwrappedcauchy}
\title{Density, distribution, random number generation and quantiles of the wrapped
Cauchy distribution}
\usage{
rwrappedcauchy(n, location = 0, scale = 1, ...)

dwrappedcauchy(theta, location = 0, scale = 1, K = 100)

pwrappedcauchy(theta, location = 0, scale = 1, K = 100)

qwrappedcauchy(p, location = 0, scale = 1, K = 100)
}
\arguments{
\item{n}{Number of random samples to be generated.}

\item{location}{The numeric value of the mean of the distribution.}

\item{scale}{The numeric value of the parameter tuning the spread of the
density.}

\item{...}{Additional arguments (NOT USED ???)}

\item{theta}{Anumeric vector hoding the angles at which to calculate the density or distribution.}

\item{K}{The limit used to approximate the distribution.}

\item{p}{vector of probabilities}
}
\description{
Distribution function and quantiles of the wrapped Cauchy distribution can
not be obtained analytically. They are therefore missing in the
\code{circular} package and are obtained here numerically.
\code{rwrappedcauchy()} and \code{dwrappedcauchy()} don't need a numerical
approximation and are just here for consistency in parametrization with the
other wrappedcauchy functions. You could also convert scale to rho
(rho=exp(-scale)) and use \code{circular::rwrappedcauchy()}. In fact, for
\code{dwrappedcauchy()}, you usually SHOULD convert scale to rho
(rho=exp(-scale)) and use \code{circular::dwrappedcauchy()} which does not
make a numerical approximation and is faster than
\code{cylcop::dwrappedcauchy()}.
}
